Moves redundancies to new Trait.
