+++ Integration/VirtualFilesystemTestCase.php
@@ -3,32 +3,33 @@
 namespace WPMedia\PHPUnit\Integration;
 
 use Brains\Monkey\Functions;
-use WPMedia\PHPUnit\VirtualFilesystemDirect;
+use WPMedia\PHPUnit\VirtualFilesystemTestTrait;
 
 abstract class VirtualFilesystemTestCase extends TestCase {
+
+	use VirtualFilesystemTestTrait;
+
+	/**
+	 * Path to the Fixtures directory.
+	 *
+	 * @var string
+	 */
+	protected static $path_to_fixtures_dir;
+
+	/**
+	 * Path to the config and test data in the Fixtures directory.
+	 * Set this path in each test class.
+	 *
+	 * @var string
+	 */
+	protected $path_to_test_data;
 
 	/**
 	 * The root virtual directory name.
 	 *
 	 * @var string
 	 */
-	protected $rootVirtualDir = 'cache';
-
-	/**
-	 * Virtual Directory filesystem structure under the root virtual directory.
-	 *
-	 * @var array
-	 */
-	protected $structure = [
-		'busting'      => [
-			'1' => [],
-		],
-		'critical-css' => [],
-		'min'          => [],
-		'wp-rocket'    => [
-			'index.html' => '',
-		],
-	];
+	protected $rootVirtualDir = 'public';
 
 	/**
 	 * Virtual directory and file permissions.
@@ -38,26 +39,11 @@ abstract class VirtualFilesystemTestCase extends TestCase {
 	protected $permissions = 0777;
 
 	/**
-	 * Instance of the virtual filesystem.
-	 *
-	 * @var VirtualFilesystemDirect
-	 */
-	protected $filesystem;
-
-	/**
-	 * URL to the root directory of the virtual filesystem.
-	 *
-	 * @var string
-	 */
-	protected $rootVirtualUrl;
-
-	/**
 	 * Prepares the test environment before each test.
 	 */
 	public function setUp() {
-		parent::setUp();
+		$this->init();
 
-		$this->filesystem     = new VirtualFilesystemDirect( $this->rootVirtualDir, $this->structure, $this->permissions );
-		$this->rootVirtualUrl = $this->filesystem->getUrl( $this->rootVirtualDir );
+		parent::setUp();
 	}
 }
